requirements.txt
beautifulsoup4==4.10.0
grequests==0.6.0
PyYAML==6.0
requests==2.25.1
